Status Change Available
Mobility Data Specification/Status Change
The view contains a query/report that shows information of the 
Vehicles available for customer use. (Status Change-When it was available and when it stopped becoming available). 
The attached file contains 2 parts. 
A create table statement to create a table and the View containing the query.


The sample data set and the insert statement is not included in this doc.

The View/query/results contains: 

The set of results, CTE (common table expression) contains a select query that creates a row number column with sorted fields. 
The CTE as table B is joined to table A to create a row-record to show a side by side comparison between the field/values from table A and B(CTE). 

CTE as table B is joined to table A and join each row to the next row after using +1
Filter: Where clause shows for the event type field all the available in table A and anything other than available in table B